refactor: state용 패키지 수정
