+++ src/main/java/blackjack/domain/state/AbstractBlackjackGameState.java
@@ -7,7 +7,7 @@ import java.util.Objects;
 
 public abstract class AbstractBlackjackGameState implements BlackjackGameState {
 
-    final Cards cards;
+    protected final Cards cards;
 
     public AbstractBlackjackGameState(final Cards cards) {
         Objects.requireNonNull(cards, "cards는 null이 들어올 수 없습니다.");
