+++ src/main/java/blackjack/domain/state/finished/Blackjack.java
@@ -1,13 +1,14 @@
-package blackjack.domain.state;
+package blackjack.domain.state.finished;
 
 import blackjack.domain.card.Cards;
+import blackjack.domain.state.BlackjackGameState;
 
 public final class Blackjack extends Finished {
 
     private static final double BLACKJACK_WIN_RATE = 1.5;
     private static final double BLACKJACK_DRAW_RATE = 0;
 
-    Blackjack(final Cards cards) {
+    public Blackjack(final Cards cards) {
         super(cards, cards.score());
     }
