+++ src/main/java/blackjack/domain/state/finished/Bust.java
@@ -1,13 +1,14 @@
-package blackjack.domain.state;
+package blackjack.domain.state.finished;
 
 import blackjack.domain.card.Cards;
 import blackjack.domain.card.Score;
+import blackjack.domain.state.BlackjackGameState;
 
 public final class Bust extends Finished {
 
     private static final double BUST_LOSE_RATE = -1;
 
-    Bust(final Cards cards, final Score score) {
+    public Bust(final Cards cards, final Score score) {
         super(cards, score);
     }
