+++ src/main/java/blackjack/domain/state/running/DealerRunning.java
@@ -1,7 +1,10 @@
-package blackjack.domain.state;
+package blackjack.domain.state.running;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
+import blackjack.domain.state.BlackjackGameState;
+import blackjack.domain.state.finished.Bust;
+import blackjack.domain.state.finished.Stand;
 
 public final class DealerRunning extends Running {
