+++ src/test/java/blackjack/domain/state/running/DealerRunningTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.state;
+package blackjack.domain.state.running;
 
 import static blackjack.domain.card.Denomination.A;
 import static blackjack.domain.card.Denomination.EIGHT;
@@ -12,6 +12,9 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
+import blackjack.domain.state.BlackjackGameState;
+import blackjack.domain.state.finished.Bust;
+import blackjack.domain.state.finished.Stand;
 import java.util.List;
 import org.junit.jupiter.api.Test;
